4_execs-and-directors/requirements.txt
crewai
langchain-google-genai
ddgs
pypdf
httpx
python-dotenv
pydantic>=2.0.0
